Make walls impassable + map::Map refactor
Now a Map has a standalone object of its own.

+++ src/consts.rs
@@ -1,2 +1,2 @@
-pub const SCREEN_WIDTH: u32 = 80;
-pub const SCREEN_HEIGHT: u32 = 50;
+pub const SCREEN_WIDTH: usize = 80;
+pub const SCREEN_HEIGHT: usize = 50;
